new file structure of project
