+++ src/bot/commands/Command.ts
