+++ src/bot/config/ConfigService.interface.ts
