--- src/constants.ts
@@ -1,16 +1,0 @@
-// Statuses for logging
-export const statuses = {
-  NEW_PARTICIPANT: 'New participant',
-  PARTICIPANT_UPDATE: 'Participant update',
-  EVENT_UPDATE: 'Event update',
-};
-
-// General messages from database
-export const messages = {
-  START_MESSAGES: 'StartMessages',
-  PARTICIPANT_MESSAGES: 'ParticipantMessages',
-  VOLUNTEER_MESSAGES: 'VolunteerMessages',
-  ORGANIZER_MESSAGES: 'OrganizerMessages',
-  SPEAKER_MESSAGES: 'SpeakerMessages',
-  SPONSOR_MESSAGES: 'SponsorMessages',
-};
