+++ src/data/logger/constants.ts
@@ -1,0 +1,11 @@
+export const loggerColors = {
+  error: 'red',
+  warn: 'yellow',
+  info: 'blue',
+};
+
+export const loggerLevels = {
+  error: 'error',
+  warn: 'warn',
+  info: 'info',
+};
