--- src/logger/constants.ts
@@ -1,11 +1,0 @@
-export const loggerColors = {
-  error: 'red',
-  warn: 'yellow',
-  info: 'blue',
-};
-
-export const loggerLevels = {
-  error: 'error',
-  warn: 'warn',
-  info: 'info',
-};
