+++ src/utils/sendMessage.ts
@@ -1,13 +1,13 @@
 import { InlineKeyboardButton } from 'telegraf/typings/core/types/typegram';
 import { Markup } from 'telegraf';
 import { ObjectId } from 'mongodb';
-import { Media, Message } from '../types';
-import TelegramBot, { BotContext } from '../TelegramBot';
+import { Media, Message } from '../data/types';
+import TelegramBot, { BotContext } from '../bot/TelegramBot';
 import 'dotenv/config';
 import parseRichText from './parseRichText';
 import { isValidUrl } from './isValidUrl';
 import getErrorMsg from './getErrorMessage';
-import logger from '../logger/logger';
+import logger from '../data/logger/logger';
 
 /**
  * Send a message with images and/or buttons from Standard Messages DB collection.
